+++ src/App.js
@@ -15,9 +15,9 @@ import {
 import './App.css';
 import Footer from './Components/Footer/Footer';
 
-import Splash from './Components/Splash';
+import Splash from './Components/Splash/Splash';
 import approvalIcon from './Assets/Images/trans-flag-graphic.svg';
-import AdminView from './Components/AdminView';
+import AdminView from './Components/Dashboard/AdminView';
 import NavigatorMenu from './Components/Navigation/NavigatorMenu';
 import MilestoneMap from './Components/Milestones/MilestoneMap';
