+++ src/Components/Dashboard/AvatarCard.js
@@ -10,7 +10,7 @@ import {
   doc, setDoc, deleteDoc,
 } from 'firebase/firestore';
 import { PropTypes } from 'prop-types';
-import { db } from './firebase';
+import { db } from '../../firebase';
 import './AvatarCard.css';
 
 const options = [
