+++ src/Pages/Profile.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ProfileOutline from '../Components/profileOutline';
+import ProfileOutline from '../Components/Profile/Profile';
 
 export default function Profile() {
   return (
