+++ src/firebase.js
@@ -1,7 +1,19 @@
 // Import the functions you need from the SDKs you need
 import { initializeApp } from 'firebase/app';
+import {
+  getAuth,
+  updateProfile,
+  signInWithEmailAndPassword,
+  createUserWithEmailAndPassword,
+  sendPasswordResetEmail,
+  signOut,
+} from 'firebase/auth';
+import {
+  getFirestore,
+  collection,
+  addDoc,
+} from 'firebase/firestore';
 
-import { getFirestore } from 'firebase/firestore';
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
 
@@ -23,6 +35,71 @@ const firebaseConfig = {
 
 // Initialize Firebase
 const app = initializeApp(firebaseConfig);
+const auth = getAuth(app);
+const db = getFirestore(app);
+
+const logInWithEmailAndPassword = async (email, password) => {
+  try {
+    await signInWithEmailAndPassword(auth, email, password);
+  } catch (err) {
+    console.error(err);
+    alert(err.message);
+  }
+  return true;
+};
+
+const registerWithEmailAndPassword = async (
+  firstName,
+  lastName,
+  accountType,
+  email,
+  password,
+  setDisplayName,
+) => {
+  try {
+    const res = await createUserWithEmailAndPassword(auth, email, password);
+    const { user } = res;
+    setDisplayName(firstName);
+    await addDoc(collection(db, 'allUsers'), {
+      uid: user.uid,
+      firstName,
+      lastName,
+      accountType,
+      authProvider: 'local',
+      email,
+    }).then(() => {
+      updateProfile(auth.currentUser, {
+        displayName: firstName,
+      });
+    });
+  } catch (err) {
+    console.error(err);
+    alert(err.message);
+  }
+  return true;
+};
+
+const sendPasswordReset = async (email) => {
+  try {
+    await sendPasswordResetEmail(auth, email);
+    alert('Password reset link sent!');
+  } catch (err) {
+    console.error(err);
+    alert(err.message);
+  }
+};
+
+const logout = () => {
+  signOut(auth);
+};
 
 // export const db = getFirestore(app);
 export default getFirestore(app);
+export {
+  auth,
+  db,
+  logInWithEmailAndPassword,
+  registerWithEmailAndPassword,
+  sendPasswordReset,
+  logout,
+};
